refactor: UserResultTable.js를 index.js로 통합 변경
