+++ src/components/containers/UserResult/index.js
@@ -1,8 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { UserResultTable } from './UserResultTable';
-import { Animation, Button, Record, Title, XButton } from '../../shared';
-import { getComputedResult } from './service';
+import {
+  Animation,
+  Button,
+  Record,
+  Title,
+  XButton,
+  LottoBall,
+  Table,
+  Thead,
+  Tbody,
+  TbodyRow,
+} from '../../shared';
+import { getComputedResult, getNumOfMatch } from './service';
+import { RESULT_TABLE_DATA } from '../../../constants';
 import { coin } from '../../../statics';
 import './style.css';
 
@@ -60,4 +71,44 @@ UserResult.propTypes = {
   lottoBundle: PropTypes.array.isRequired,
   winningNumber: PropTypes.object.isRequired,
   onReset: PropTypes.func.isRequired,
+};
+
+function UserResultTable(props) {
+  const { lottoBundle, winningNumber } = props;
+  const { winningNumbers, bonusNumber } = winningNumber;
+
+  const theadItems = ['구분', '번호'];
+  const getFirstCell = (lotto) => {
+    const numOfMatch = getNumOfMatch(lotto, winningNumber);
+    return RESULT_TABLE_DATA[numOfMatch].DESCRIPTION;
+  };
+  const LottoBalls = (lotto, rowIndex) =>
+    lotto.map((num, index) => (
+      <LottoBall
+        key={`${rowIndex}th-lotto-ball-${index}`}
+        targetNumber={num}
+        winningNumbers={winningNumbers.concat(bonusNumber)}
+      />
+    ));
+
+  return (
+    <Table>
+      <Thead>{theadItems}</Thead>
+      <Tbody>
+        {lottoBundle.map((lotto, index) => (
+          <TbodyRow key={index} rowIndex={index}>
+            {[getFirstCell(lotto), LottoBalls(lotto, index)]}
+          </TbodyRow>
+        ))}
+      </Tbody>
+    </Table>
+  );
+}
+
+UserResultTable.propTypes = {
+  lottoBundle: PropTypes.array.isRequired,
+  winningNumber: PropTypes.exact({
+    winningNumbers: PropTypes.array,
+    bonusNumber: PropTypes.number,
+  }).isRequired,
 };
